Split off local visibility analysis and put it (along with global) into another folder
