+++ salalib/vgamodules/vgamodules.pri
@@ -1,0 +1,4 @@
+HEADERS += vgamodules/visualglobal.h \
+    $$PWD/visuallocal.h
+SOURCES += vgamodules/visualglobal.cpp \
+    $$PWD/visuallocal.cpp
