+++ salalib/vgamodules/visualglobal.h
